Refactor Permission status (move to separate component)
