+++ ui/components/app/permission-cell/permission-cell.js
@@ -3,12 +3,10 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import {
   AlignItems,
-  BorderColor,
   Color,
   IconColor,
   JustifyContent,
   Size,
-  TextAlign,
   TextColor,
   TextVariant,
   Display,
@@ -24,16 +22,10 @@ import {
   IconSize,
   Text,
   Box,
-  AvatarTokenSize,
-  AvatarAccount,
-  AvatarAccountSize,
 } from '../../component-library';
-import { formatDate } from '../../../helpers/utils/util';
-import { useI18nContext } from '../../../hooks/useI18nContext';
 import Tooltip from '../../ui/tooltip';
-import { AvatarGroup } from '../../multichain';
-import { AvatarType } from '../../multichain/avatar-group/avatar-group.types';
 import { PermissionCellOptions } from './permission-cell-options';
+import { PermissionCellStatus } from './permission-cell-status';
 
 const PermissionCell = ({
   snapId,
@@ -48,8 +40,6 @@ const PermissionCell = ({
   hideStatus,
   accounts,
 }) => {
-  const t = useI18nContext();
-
   const infoIcon = IconName.Info;
   let infoIconColor = IconColor.iconMuted;
   let iconColor = IconColor.primaryDefault;
@@ -74,76 +64,6 @@ const PermissionCell = ({
   let permissionIcon = avatarIcon;
   if (typeof avatarIcon !== 'string' && avatarIcon?.props?.iconName) {
     permissionIcon = avatarIcon.props.iconName;
-  }
-
-  const renderAccountsGroup = () => (
-    <Box
-      as="span"
-      className="permission-cell__accounts-group-box"
-      display={Display.InlineFlex}
-    >
-      <Tooltip
-        position="bottom"
-        html={
-          <Box
-            display={Display.Flex}
-            flexDirection={FlexDirection.Column}
-            justifyContent={JustifyContent.center}
-            alignItems={AlignItems.center}
-          >
-            <Text
-              variant={TextVariant.headingSm}
-              color={TextColor.textAlternative}
-              textAlign={TextAlign.Center}
-            >
-              {t('accounts')}
-            </Text>
-            <Box display={Display.Flex} flexDirection={FlexDirection.Column}>
-              {accounts.map((account, index) => (
-                <Box
-                  key={`${account.avatarValue}_${index}`}
-                  display={Display.Flex}
-                  justifyContent={JustifyContent.flexStart}
-                  alignItems={AlignItems.center}
-                  marginTop={2}
-                >
-                  <AvatarAccount
-                    address={account.avatarValue}
-                    size={AvatarAccountSize.Xs}
-                    borderColor={BorderColor.backgroundDefault}
-                  />
-                  <Text variant={TextVariant.bodyMdMedium} marginLeft={2}>
-                    {account.avatarName}
-                  </Text>
-                </Box>
-              ))}
-            </Box>
-          </Box>
-        }
-      >
-        <AvatarGroup
-          limit={3}
-          members={accounts}
-          avatarType={AvatarType.ACCOUNT}
-          size={AvatarTokenSize.Xs}
-          width={BlockSize.Min}
-          borderColor={BorderColor.backgroundDefault}
-          marginLeft={4}
-          paddingLeft={4}
-        />
-      </Tooltip>
-    </Box>
-  );
-
-  let accountGroup = '';
-  let statusRequestedTranslation = 'permissionRequested';
-  let statusApprovedTranslation = 'approvedOn';
-  let statusRevokedTranslation = 'permissionRevoked';
-  if (accounts && accounts.length) {
-    accountGroup = renderAccountsGroup();
-    statusRequestedTranslation = 'permissionRequestedForAccounts';
-    statusApprovedTranslation = 'approvedOnForAccounts';
-    statusRevokedTranslation = 'permissionRevokedForAccounts';
   }
 
   return (
@@ -188,21 +108,11 @@ const PermissionCell = ({
           {title}
         </Text>
         {!hideStatus && (
-          <Text
-            className="permission-cell__status"
-            variant={TextVariant.bodySm}
-            color={TextColor.textAlternative}
-            display={Display.Flex}
-          >
-            {!revoked &&
-              (dateApproved
-                ? t(statusApprovedTranslation, [
-                    formatDate(dateApproved, 'yyyy-MM-dd'),
-                    accountGroup,
-                  ])
-                : t(statusRequestedTranslation, [accountGroup]))}
-            {revoked ? t(statusRevokedTranslation, [accountGroup]) : ''}
-          </Text>
+          <PermissionCellStatus
+            revoked={revoked}
+            dateApproved={dateApproved}
+            accounts={accounts}
+          />
         )}
       </Box>
       <Box display={Display.Flex}>
